requirements.txt
opencv-contrib-python
mysql-connector-python
pandas
numpy
Pillow
